requirements.txt
tensorflow-gpu==2.9
tensorflow-addons
tensorflow-hub
tensorflow-datasets
graphlearning
pydub
numpy